chore: deprecate legacy code
