+++ frontend/src/components/SensitiveData/SensitiveColumnDrawer.vue
@@ -172,12 +172,12 @@ import { NSelect, NButton } from "naive-ui";
 import { computed, reactive, onMounted } from "vue";
 import { useI18n } from "vue-i18n";
 import { updateColumnConfig } from "@/components/ColumnDataTable/utils";
+import type { MaskData } from "@/components/SensitiveData/types";
 import { useSemanticType } from "@/components/SensitiveData/useSemanticType";
 import { Drawer, DrawerContent } from "@/components/v2";
 import { useSettingV1Store, useDBSchemaV1Store } from "@/store";
 import { type ComposedDatabase } from "@/types";
 import { MaskingLevel } from "@/types/proto/v1/common";
-import type { MaskData } from "@/types/proto/v1/org_policy_service";
 import { hasWorkspacePermissionV2 } from "@/utils";
 import GrantAccessDrawer from "./GrantAccessDrawer.vue";
 import MaskingExceptionUserTable from "./MaskingExceptionUserTable.vue";
