+++ frontend/src/components/SensitiveData/types.ts
@@ -2,10 +2,16 @@ import type { ComposedDatabase, DatabaseResource } from "@/types";
 import { type User } from "@/types/proto/v1/auth_service";
 import { MaskingLevel } from "@/types/proto/v1/common";
 import type { Group } from "@/types/proto/v1/group_service";
-import type {
-  MaskData,
-  MaskingExceptionPolicy_MaskingException_Action,
-} from "@/types/proto/v1/org_policy_service";
+import type { MaskingExceptionPolicy_MaskingException_Action } from "@/types/proto/v1/org_policy_service";
+
+export interface MaskData {
+  schema: string;
+  table: string;
+  column: string;
+  maskingLevel: MaskingLevel;
+  fullMaskingAlgorithmId: string;
+  partialMaskingAlgorithmId: string;
+}
 
 export interface SensitiveColumn {
   database: ComposedDatabase;
